chore: Create necessary directories and implement file locking mechanism for pptx to pdf conversion script

+++ shellscripts/pptx_to_pdf.sh
@@ -7,12 +7,13 @@
 PDF_FOLDER="$HOME/Documents/10_PDF檔" # PDF 存放目錄
 SOURCE_FOLDER="${PDF_FOLDER}"        # 使用 {} 進行變數替換
 CONVERTED_FOLDER="${PDF_FOLDER}/converted"
-CCO_SLIDES_DIR="${PDF_FOLDER}/CCO_slides"
 LIBREOFFICE="/Applications/LibreOffice.app/Contents/MacOS/soffice"
+LOCK_DIR="$PDF_FOLDER/locks" # 鎖檔目錄
 
-# 檢查並創建 converted 和 PDF 存放目錄
+# 檢查並創建必要目錄
 mkdir -p "$PDF_FOLDER"
 mkdir -p "$CONVERTED_FOLDER"
+mkdir -p "$LOCK_DIR"
 
 # 尋找當前目錄中的 .ppt 和 .pptx 檔案
 FILES=$(find "$SOURCE_FOLDER" -maxdepth 1 -type f \( -name "*.ppt" -o -name "*.pptx" \))
@@ -23,22 +24,31 @@ if [ -z "$FILES" ]; then
 else
   # 如果找到檔案，則進行轉換
   echo "$FILES" | while read file; do
-    # 轉換檔案為 PDF
-    "$LIBREOFFICE" --headless --convert-to pdf --outdir "$PDF_FOLDER" "$file"
-    echo "正準備開始轉換\033[0;32m$file\033[0m"
-    # 轉換後移動原始檔案到 converted 目錄
-    #
-    # 獲取檔案基本名稱（不含路徑）
     base_name=$(basename "$file")
+    lock_file="$LOCK_DIR/${base_name}.lock"
 
-    # 如果檔案名稱以 "CCO" 開頭，則移動 PDF 到 CCO_slides 目錄
-    # if [[ $base_name == CCO* ]]; then
-    # 	# 構建 PDF 檔案名稱
-    # 	pdf_file="${PDF_FOLDER}/${base_name%.*}.pdf"
-    # 	mv "$pdf_file" "$CCO_SLIDES_DIR/"
-    # fi
+    # 檢查是否有鎖檔案
+    if [ -e "$lock_file" ]; then
+      echo "\033[0;33m檔案 $file 已被另一個進程處理，跳過。\033[0m"
+      continue
+    fi
+
+    # 創建鎖檔案
+    touch "$lock_file"
+
+    # 顯示通知：發現檔案並開始轉換
+    osascript -e "display notification \"開始轉換: $base_name\" with title \"發現檔案\""
+
+    # 轉換檔案為 PDF
+    echo "正準備開始轉換\033[0;32m$file\033[0m"
+    "$LIBREOFFICE" --headless --convert-to pdf --outdir "$PDF_FOLDER" "$file"
+
+    # 移動原始檔案到 converted 目錄
     mv "$file" "$CONVERTED_FOLDER/"
     echo "\033[0;32m轉換完成\033[0m"
     osascript -e "display notification \"轉換完成\" with title \"$base_name\""
+
+    # 刪除鎖檔案
+    rm -f "$lock_file"
   done
 fi
